requirements.txt
discord.py
python-dotenv
pandas